Remove unnecessary UTF-8 decoding step during PDF output.
This adds a new variant `PandocOutput::ToBufferRaw`, allowing us to
return to the user the raw bytes given to us by pandoc in the case of
binary output formats (for now this is just PDF).

This is a breaking change since a new variant is introduced to an
existing enum.