Adds ResourcePath PandocOption.
When generating outputs that embed resources such as images (this
includes PDF, ePub, amongst other formats), typically pandoc looks in
the working directory to find these files. Using the `--resource-path`
option, which was introduced in recent versions of Pandoc (2.10), this
behhaviour can be changed.

With this commit, multiple resource paths can be specified. As the
pandoc manual states, these are chained with a colon (':') on most
operating systems, but with a semicolon (';') on the 'Windows' operating
system.